requirements.txt
flask
gunicorn
langchain==0.3.25
langchain_core==0.3.63
pandas==2.2.2
sentence_transformers==4.1.0
torch==2.7.0
transformers==4.52.3
pyyaml